+++ admin/class-decalog-admin.php
@@ -537,6 +537,7 @@ class Decalog_Admin {
 				Option::network_set( 'use_cdn', array_key_exists( 'decalog_plugin_options_usecdn', $_POST ) );
 				Option::network_set( 'display_nag', array_key_exists( 'decalog_plugin_options_nag', $_POST ) );
 				Option::network_set( 'download_favicons', array_key_exists( 'decalog_plugin_options_favicons', $_POST ) ? (bool) filter_input( INPUT_POST, 'decalog_plugin_options_favicons' ) : false );
+				Option::network_set( 'earlyloading', array_key_exists( 'decalog_plugin_options_earlyloading', $_POST ) ? (bool) filter_input( INPUT_POST, 'decalog_plugin_options_earlyloading' ) : false );
 				Option::network_set( 'logger_autostart', array_key_exists( 'decalog_loggers_options_autostart', $_POST ) );
 				Option::network_set( 'pseudonymization', array_key_exists( 'decalog_loggers_options_pseudonymization', $_POST ) );
 				Option::network_set( 'respect_wp_debug', array_key_exists( 'decalog_loggers_options_wpdebug', $_POST ) );
@@ -810,6 +811,22 @@ class Decalog_Admin {
 	 */
 	public function plugin_options_section_callback() {
 		$form = new Form();
+		add_settings_field(
+			'decalog_plugin_options_earlyloading',
+			__( 'Initialization', 'decalog' ),
+			[ $form, 'echo_field_checkbox' ],
+			'decalog_plugin_options_section',
+			'decalog_plugin_options_section',
+			[
+				'text'        => esc_html__( 'Activate early loading', 'decalog' ),
+				'id'          => 'decalog_plugin_options_earlyloading',
+				'checked'     => Option::network_get( 'earlyloading' ),
+				'description' => esc_html__( 'If checked, DecaLog will be loaded before all other plugins (recommended).', 'decalog' ),
+				'full_width'  => false,
+				'enabled'     => true,
+			]
+		);
+		register_setting( 'decalog_plugin_options_section', 'decalog_plugin_options_earlyloading' );
 		add_settings_field(
 			'decalog_plugin_options_favicons',
 			__( 'Favicons', 'decalog' ),